src/predictionModel/app/requirements.txt
numpy==1.19.4
pandas==1.2.1
scikit-learn ==0.23.2
torch==1.8.1
tqdm==4.50.2
numpy==1.19.4
pickleshare==0.7.5
requests==2.25.0
flask==1.1.2
google-cloud-bigquery[bqstorage,pandas]
google-cloud-storage